refactor: decompose NpcManager god object into focused modules (closes #697)
Extract four subsystems out of the 2,633-line manager.rs so that each
concern — schedule resolution, tier assignment, banshee, Tier 4 event
application — lives in one file with its types, logic, and tests.

New modules:
- schedule.rs — ScheduleEvent/ScheduleEventKind + tick_schedules free fn
- tier_assign.rs — TierTransition + assign_tiers + BFS helpers

Augmented modules:
- banshee.rs — absorbs tick_banshee loop (banshee::tick)
- tier4.rs — absorbs apply_tier4_events (tier4::apply_events)

NpcManager in manager.rs is now a coordinator: struct + CRUD +
introduction tracking + tick state management + thin wrapper methods
that delegate to the subsystem free functions. Types ScheduleEvent,
ScheduleEventKind, and TierTransition are re-exported from manager so
existing callers need no import changes.

Tests follow the code they test. Shared test fixtures (make_test_npc,
make_scheduled_npc, make_chain_graph, etc.) consolidated into a single
used by all test modules.
